[SC test & script]: Replace the uint256[] with the uint8[] in the InsuranceClaimProofVerifier.t.sol and PubkeyAndSignedMessageExtractor.sol

+++ scripts/utils/array-bytes-generator/contracts/libraries/PubkeyAndSignedMessageExtractor.sol
@@ -12,10 +12,14 @@ contract PubkeyAndSignedMessageExtractor is Script {
      * @dev - Struct to store the PubkeyAndSignedMessage related data (to be the bytes64 Field data)
      */
     struct PubkeyAndSignedMessage {
-        uint256[] insurer_signature_bytes;
-        uint256[] insurer_pubkey_bytes;
-        uint256[] hospital_pubkey_bytes;
-        uint256[] hospital_signature_bytes;
+        uint8[] insurer_signature_bytes;
+        uint8[] insurer_pubkey_bytes;
+        uint8[] hospital_pubkey_bytes;
+        uint8[] hospital_signature_bytes;
+        // uint256[] insurer_signature_bytes;
+        // uint256[] insurer_pubkey_bytes;
+        // uint256[] hospital_pubkey_bytes;
+        // uint256[] hospital_signature_bytes;
     }
 
     /**
@@ -36,33 +40,41 @@ contract PubkeyAndSignedMessageExtractor is Script {
         //console.logBytes(data);
 
         /// @dev - Define the uint256 array variables (64 bytes each)
-        uint256[] memory insurer_pubkey_bytes_array = new uint256[](64);
-        uint256[] memory insurer_signature_bytes_array = new uint256[](64);
-        uint256[] memory hospital_pubkey_bytes_array = new uint256[](64);
-        uint256[] memory hospital_signature_bytes_array = new uint256[](64);
+        uint8[] memory insurer_pubkey_bytes_array = new uint8[](64);
+        uint8[] memory insurer_signature_bytes_array = new uint8[](64);
+        uint8[] memory hospital_pubkey_bytes_array = new uint8[](64);
+        uint8[] memory hospital_signature_bytes_array = new uint8[](64);
+        // uint256[] memory insurer_pubkey_bytes_array = new uint256[](64);
+        // uint256[] memory insurer_signature_bytes_array = new uint256[](64);
+        // uint256[] memory hospital_pubkey_bytes_array = new uint256[](64);
+        // uint256[] memory hospital_signature_bytes_array = new uint256[](64);
         console.log("insurer_pubkey_bytes_array.length: %s", insurer_pubkey_bytes_array.length); // [Log]: 64
 
         /// @dev - Store a Uint8Array value, which was retrieved from the output.json, into the uint256 array variable (uint256[])
-        uint256[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
+        uint8[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
+        //uint256[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
         console.log("_insurer_signature_bytes.length: %s", _insurer_signature_bytes.length);     // [Log]: 99
         for (uint i = 0; i < _insurer_signature_bytes.length - 1; i++) {
             insurer_signature_bytes_array[i] = _insurer_signature_bytes[i];
             console.log("insurer_signature_bytes_array[%s] = %s", i, insurer_signature_bytes_array[i]); // [Log - Success]: _insurer_signature_bytes[0] = 211, _insurer_signature_bytes[1] = 23, ...
         }
 
-        uint256[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
+        uint8[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
+        //uint256[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
         for (uint i = 0; i < _hospital_signature_bytes.length - 1; i++) {
             hospital_signature_bytes_array[i] = _hospital_signature_bytes[i];
             console.log("hospital_signature_bytes_array[%s] = %s", i, hospital_signature_bytes_array[i]); // [Log - Success]: _hospital_signature_bytes[0] = 211, _hospital_signature_bytes[1] = 23, ...
         }
 
-        uint256[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
+        uint8[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
+        //uint256[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
         for (uint i = 0; i < _insurer_pubkey_bytes.length; i++) {
             insurer_pubkey_bytes_array[i] = _insurer_pubkey_bytes[i];
             console.log("insurer_pubkey_bytes_array[%s] = %s", i, insurer_pubkey_bytes_array[i]); // [Log - Success]: _insurer_pubkey_bytes[0] = 0x215597bacd9c7e977dfc170f320074155de974be494579d2586e5b268fa3b629, _insurer_pubkey_bytes[1] = 0x26df0d347e961cb94e1cc6d2ad8558696de8c1964b30e26f2ec8b926cbbbf862, ...
         }
 
-        uint256[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
+        uint8[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
+        //uint256[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
         for (uint i = 0; i < _hospital_pubkey_bytes.length - 1; i++) {
             hospital_pubkey_bytes_array[i] = _hospital_pubkey_bytes[i];
             console.log("hospital_signature_bytes_array[%s] = %s", i, hospital_signature_bytes_array[i]); // [Log - Success]: _hospital_pubkey_bytes[0] = 0x215597bacd9c7e977dfc170f320074155de974be494579d2586e5b268fa3b629, _hospital_pubkey_bytes[1] = 0x26df0d347e961cb94e1cc6d2ad8558696de8c1964b30e26f2ec8b926cbbbf862, ...
